COMMERCE-766 Move to the string related operations to an util class
